Module bellpepper::gadgets

Self-contained sub-circuit implementations for various primitives.

Modules

blake2s

The BLAKE2s hash function with personalization support.

boolean

Gadgets for allocating bits in the circuit and performing boolean logic.

lookup

Window table lookup gadgets.

multieq

multipack

Helpers for packing vectors of bits into scalar field elements.

num

Gadgets representing numbers in the scalar field of the underlying curve.

sha256

Circuits for the SHA-256 hash function and its internal compression function.

uint32

Circuit representation of a u32, with helpers for the [sha256] gadgets.

Traits

Assignment

This basically is just an extension to Option which allows for a convenient mapping to an error on None.