Struct bellpepper::gadgets::boolean::AllocatedBit

pub struct AllocatedBit { /* private fields */ }

Represents a variable in the constraint system which is guaranteed to be either zero or one.

Implementations

impl AllocatedBit

pub fn get_value(&self) -> Option<bool>

pub fn get_variable(&self) -> Variable

pub fn alloc_conditionally<Scalar, CS>( cs: CS, value: Option<bool>, must_be_false: &AllocatedBit ) -> Result<AllocatedBit, SynthesisError>
where Scalar: PrimeField, CS: ConstraintSystem<Scalar>,

Allocate a variable in the constraint system which can only be a boolean value. Further, constrain that the boolean is false unless the condition is false.

pub fn alloc<Scalar, CS>( cs: CS, value: Option<bool> ) -> Result<AllocatedBit, SynthesisError>
where Scalar: PrimeField, CS: ConstraintSystem<Scalar>,

Allocate a variable in the constraint system which can only be a boolean value.

pub fn xor<Scalar, CS>( cs: CS, a: &AllocatedBit, b: &AllocatedBit ) -> Result<AllocatedBit, SynthesisError>
where Scalar: PrimeField, CS: ConstraintSystem<Scalar>,

Performs an XOR operation over the two operands, returning an AllocatedBit.

pub fn and<Scalar, CS>( cs: CS, a: &AllocatedBit, b: &AllocatedBit ) -> Result<AllocatedBit, SynthesisError>
where Scalar: PrimeField, CS: ConstraintSystem<Scalar>,

Performs an AND operation over the two operands, returning an AllocatedBit.

pub fn and_not<Scalar, CS>( cs: CS, a: &AllocatedBit, b: &AllocatedBit ) -> Result<AllocatedBit, SynthesisError>
where Scalar: PrimeField, CS: ConstraintSystem<Scalar>,

Calculates a AND (NOT b).

pub fn nor<Scalar, CS>( cs: CS, a: &AllocatedBit, b: &AllocatedBit ) -> Result<AllocatedBit, SynthesisError>
where Scalar: PrimeField, CS: ConstraintSystem<Scalar>,

Calculates (NOT a) AND (NOT b).